Add files via upload
Implemented a Python program to encrypt and decrypt a word using a 4-letter word using hill cypher. Utilized numpy module in Python
to implement matrix calculations such as addition, multiplication, and inverse. Encryption was performed by converting the string input
to a matrix and using the 4-letter key matrix to encrypt the text. Decryption was performed by converting the key matrix to its inverse
form and multiplying it with the encrypted word matrix.